add global variable to GE_helpers and move required_is_only_rule to utils
