+++ schematic/utils/validate_utils.py
@@ -4,6 +4,7 @@
 
 import re
 from collections.abc import Mapping
+import logging
 from typing import Pattern, Union, Iterable, Any, Optional
 from numbers import Number
 from jsonschema import validate
@@ -11,6 +12,10 @@ import numpy as np
 import pandas as pd
 from schematic.utils.io_utils import load_json
 from schematic import LOADER
+
+logger = logging.getLogger(__name__)
+
+
 
 
 def validate_schema(schema: Union[Mapping, bool]) -> None:
@@ -104,3 +109,65 @@ def iterable_to_str_list(obj: Union[str, Number, Iterable]) -> list[str]:
     # If the object is iterable and not a string, convert every element
     # to string and wrap as a list
     return [str(item) for item in obj]
+
+def required_is_only_rule(
+        rule: str, attribute: str, rule_modifiers: list[str], validation_expectation: dict[str,str],
+    ) -> bool:
+        """Need to determine if required is the only rule being set. Do this way so we dont have
+        to enforce a position for it (ie, it can only be before message and after the rule).
+        This ensures that 'required' is not treated like a real rule, in the case it is
+        accidentally combined with a rule modifier. The required rule is t
+
+        Args:
+            rule: str, the validation rule string
+            validation_expectation: dict[str, str], currently implemented expectations.
+        Returns:
+            bool, True, if required is the only rule, false if it is not.
+        """
+        # convert rule to lowercase to ensure punctuation does not throw off determination.
+        rule = rule.lower()
+
+        # If required is not in the rule, it cant be the only rule, return False
+        if "required" not in rule:
+            return False
+
+        # If the entire rule is just 'required' then it is easily determined to be the only rule
+        if rule == "required":
+            return True
+
+        # Try to find an expectation rule in the rule, if there is one there log it and
+        # continue
+        # This function is called as part of an if that is already looking for in house rules
+        # so don't worry about looking for them.
+        rule_parts = rule.split(" ")
+        for idx, rule_part in enumerate(rule_parts):
+            if rule_part in validation_expectation:
+                return False
+
+        # identify then remove all rule modifiers, all that should be left is required in the
+        # case that someone used a standard modifier with required
+        idx_to_remove = []
+        if "required" in rule_parts:
+            only_rule = True
+            for idx, rule_part in enumerate(rule_parts):
+                if rule_part in rule_modifiers:
+                    idx_to_remove.append(idx)
+
+        if idx_to_remove:
+            for idx in sorted(idx_to_remove, reverse=True):
+                del rule_parts[idx]
+
+        # In this case, rule modifiers have been added to required. This is not the expected use
+        # so log a warning, but let user proceed.
+        if rule_parts == ["required"]:
+            warning_message = " ".join(
+                [
+                    f"For Attribute: {attribute}, it looks like required was set as a single rule,"
+                    f"with modifiers attached.",
+                    f"Rule modifiers do not work in conjunction with the required validation rule.",
+                    f"Please reformat your rule.",
+                ]
+            )
+            logger.warning(warning_message)
+            return True
+
